+++ RandomSkunk.Results/Maybe{T}.cs
@@ -108,7 +108,7 @@ public readonly partial struct Maybe<T> : IResult<T>, IEquatable<Maybe<T>>
     /// <param name="error">The error for the new <c>Fail</c> result.</param>
     /// <returns>A <c>Fail</c> result with the specified error.</returns>
     [StackTraceHidden]
-    public static implicit operator Maybe<T>(Error error) => Fail(error);
+    public static implicit operator Maybe<T>(Error? error) => Fail(error);
 
     /// <summary>
     /// Indicates whether the <paramref name="left"/> parameter is equal to the <paramref name="right"/> parameter.