+++ RandomSkunk.Results/Result.cs
@@ -89,7 +89,7 @@ public readonly partial struct Result : IResult<Unit>, IEquatable<Result>
     /// <param name="error">The error for the new <c>Fail</c> result.</param>
     /// <returns>A <c>Fail</c> result with the specified error.</returns>
     [StackTraceHidden]
-    public static implicit operator Result(Error error) => Fail(error);
+    public static implicit operator Result(Error? error) => Fail(error);
 
     /// <summary>
     /// Indicates whether the <paramref name="left"/> parameter is equal to the <paramref name="right"/> parameter.